+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - license changed to MIT
 - refactoring source for core version 0.5.7
+- added module settings